먹기전, 먹은 후 전자저울 값 가져오기 ok

+++ app/src/main/java/com/example/milkbottle/GraphActivity.java
@@ -59,7 +59,7 @@ public class GraphActivity extends AppCompatActivity {
 
             for(MilkData milkDate : milkData){
                 long dateTime = milkDate.getCurrDate().getTime();
-                int quantity = milkDate.getQuantity();
+                Float quantity = milkDate.getQuantity();
                 entries.add(new Entry(dateTime, quantity));
                 dateList.add(milkDate.getCurrDate());
                 dateListF.add(milkDate.getCurrFloat());
